added user and chat route with services
